--- fixtures/app_with_git_helper_plugin/Project.swift
@@ -1,17 +1,0 @@
-import ProjectDescription
-
-let project = Project(
-    name: "App",
-    organizationName: "tuist.io",
-    targets: [
-        Target(
-            name: "App",
-            platform: .iOS,
-            product: .app,
-            bundleId: "io.tuist.app",
-            deploymentTarget: .iOS(targetVersion: "13.0", devices: .iphone),
-            infoPlist: .default,
-            sources: ["Sources/**"]
-        )
-    ]
-)
