--- fixtures/app_with_git_helper_plugin/Tuist/Config.swift
@@ -1,8 +1,0 @@
-import ProjectDescription
-
-let config = Config(
-    plugins: [
-        .git(url: "https://github.com/tuist/ExampleTuistHelpersPlugin.git", branch: "main")
-    ],
-    generationOptions: []
-)
